+++ app/Http/Controllers/TagDispersionController.php
@@ -99,7 +99,7 @@ class TagDispersionController extends Controller
         $cleanProjectId = $cleanProjectId === '' ? null : $cleanProjectId;
 
         $cleanMes = trim($request->input('mes'));
-        $cleanMes = $cleanMes === '' ? null : (float) $cleanMes;
+        $cleanMes = $cleanMes === '' ? null : (int) $cleanMes;
 
         $cleanVehicleId = trim($request->input('vehicle_id'));
         $cleanVehicleId = $cleanVehicleId === '' ? null : $cleanVehicleId;
